chore: rename StaffService directory to CaseService
