+++ api/src/middlewares/services/CaseService/__tests__/fixtures/fields/CNIAndMSC.ts
