+++ api/src/middlewares/services/CaseService/mappings/CNIAndMSC.ts
