+++ api/src/middlewares/services/CaseService/mappings/affirmation.ts
