+++ api/src/middlewares/services/CaseService/mappings/exchange.ts
