+++ api/src/middlewares/services/CaseService/utils/getApplicationTypeName.ts
