+++ api/src/middlewares/services/CaseService/utils/reorderSectionsWithNewName.ts
